CodeStyle correction

+++ src/Api/UseCase/LinkGroup/DeleteLinkGroup/DeleteLinkGroupUseCase.ts
@@ -1,7 +1,8 @@
-import { BadRequestException, Injectable } from "@nestjs/common";
-import { LinkGroupRepository } from "../../../Repository/LinkGroupRepository";
-import {ContextualGraphqlRequest, UseCase} from "../../../../index";
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
+import LinkGroupRepository from "../../../Repository/LinkGroupRepository";
+import { ContextualGraphqlRequest, UseCase } from "../../../../index";
 import { LinkGroup } from "../../../Entity/LinkGroup";
+import InsufficientPermissionException from "../../../../Core/Exception/InsufficientPermissionException";
 
 @Injectable()
 export default class DeleteLinkGroupUseCase implements UseCase<Promise<LinkGroup>,[LinkGroupId:number]> {
@@ -9,6 +10,13 @@ export default class DeleteLinkGroupUseCase implements UseCase<Promise<LinkGroup
 
     async handle(context: ContextualGraphqlRequest, linkGroupId: number): Promise<LinkGroup> {
         try {
+            const linkGroup = await this.linkGroupRepository.findById(linkGroupId);
+
+            if (!linkGroup) throw new NotFoundException(`LinkGroup with ID ${linkGroupId} not found`);
+
+            if (linkGroup.userId !== context.userId)
+                throw new InsufficientPermissionException("Access denied: User with ID ${context.userId} cannot modify link group owned by user with ID ${linkGroup.userId}.");
+
             return await this.linkGroupRepository.delete(linkGroupId, context.userId);
         } catch (error) {
             throw new BadRequestException('Failed to delete link group', error.message);
